Move SpatialIndex and LineSegment from core to utils namespace
