+++ src/core/Core.hpp
@@ -15,10 +15,8 @@
 #include "ChipmunkHelpers.hpp"
 #include "Exception.hpp"
 #include "InputDispatcher.hpp"
-#include "LineSegment.hpp"
 #include "MathHelpers.hpp"
 #include "Signals.hpp"
-#include "SpatialIndex.hpp"
 #include "StopWatch.hpp"
 #include "Strings.hpp"
 #include "TimeState.hpp"
@@ -27,9 +25,11 @@
 
 // higher level stuff
 
+#include "App.hpp"
 #include "Object.hpp"
 #include "Entity.hpp"
 #include "Stage.hpp"
+#include "Compositor.hpp"
 #include "Scenario.hpp"
 #include "Tracking.hpp"
